Moves mapping to mapper
