+++ phpunit.xml.dist
@@ -6,7 +6,7 @@
          convertNoticesToExceptions  = "true"
          convertWarningsToExceptions = "true"
          processIsolation            = "false"
-         stopOnFailure               = "true"
+         stopOnFailure               = "false"
          syntaxCheck                 = "false"
          bootstrap                   = "tests/bootstrap.php">
