+++ src/Transfer/EzPlatform/Repository/Manager/ContentManager.php
@@ -117,7 +117,7 @@ class ContentManager implements LoggerAwareInterface, CreatorInterface, UpdaterI
             $object->getProperty('language')
         );
 
-        $this->mapObjectToCreateStruct($object, $createStruct);
+        $object->getMapper()->mapObjectToCreateStruct($object, $createStruct);
 
         /** @var LocationObject[] $locationObjects */
         $locationObjects = $object->getProperty('parent_locations');
@@ -161,7 +161,7 @@ class ContentManager implements LoggerAwareInterface, CreatorInterface, UpdaterI
         $contentDraft = $this->contentService->createContentDraft($object->getProperty('content_info'));
 
         $contentUpdateStruct = $this->contentService->newContentUpdateStruct();
-        $this->mapObjectToUpdateStruct($object, $contentUpdateStruct);
+        $object->getMapper()->mapObjectToUpdateStruct($object, $contentUpdateStruct);
 
         $contentDraft = $this->contentService->updateContent($contentDraft->versionInfo, $contentUpdateStruct);
         $content = $this->contentService->publishVersion($contentDraft->versionInfo);
@@ -234,60 +234,5 @@ class ContentManager implements LoggerAwareInterface, CreatorInterface, UpdaterI
         $object->setProperty('main_location_id', $location->id);
 
         return $this->contentService->updateContentMetadata($object->getProperty('content_info'), $contentMetadataUpdateStruct);
-    }
-
-    /**
-     * Maps object data to create struct.
-     *
-     * @param ContentObject       $object       Content object to map from
-     * @param ContentCreateStruct $createStruct Content create struct to map to
-     *
-     * @throws \InvalidArgumentException
-     */
-    private function mapObjectToCreateStruct(ContentObject $object, ContentCreateStruct $createStruct)
-    {
-        if ($object->getProperty('language')) {
-            $createStruct->mainLanguageCode = $object->getProperty('language');
-        }
-
-        if ($object->getProperty('remote_id')) {
-            $createStruct->remoteId = $object->getProperty('remote_id');
-        }
-
-        $this->assignStructValues($object, $createStruct);
-    }
-
-    /**
-     * Maps object data to update struct.
-     *
-     * @param ContentObject       $object              Content object to map from
-     * @param ContentUpdateStruct $contentUpdateStruct Content update struct to map to
-     *
-     * @throws \InvalidArgumentException
-     */
-    private function mapObjectToUpdateStruct(ContentObject $object, ContentUpdateStruct $contentUpdateStruct)
-    {
-        $this->assignStructValues($object, $contentUpdateStruct);
-    }
-
-    /**
-     * Copies content object data from a struct.
-     *
-     * @param ContentObject $object Content object to get values from
-     * @param object        $struct Struct to assign values to
-     */
-    private function assignStructValues(ContentObject $object, $struct)
-    {
-        foreach ($object->data as $key => $value) {
-            if (is_array($value)) {
-                $value = end($value);
-            }
-            $struct->setField($key, $value);
-        }
-
-        if ($object->getProperty('struct_callback')) {
-            $callback = $object->getProperty('struct_callback');
-            $callback($struct);
-        }
     }
 }
